Improve folder structure
