+++ src/data/links.js
@@ -1,0 +1,27 @@
+export const links = [
+  {
+    title: 'Features',
+    sublinks: [
+      { label: 'Link Shortening', url: '/' },
+      { label: 'Branded Links', url: '/' },
+      { label: 'Branded Links', url: '/' },
+    ],
+  },
+  {
+    title: 'Resources',
+    sublinks: [
+      { label: 'Blog', url: '/' },
+      { label: 'Developers', url: '/' },
+      { label: 'Support', url: '/' },
+    ],
+  },
+  {
+    title: 'Company',
+    sublinks: [
+      { label: 'About', url: '/' },
+      { label: 'Our Team', url: '/' },
+      { label: 'Careers', url: '/' },
+      { label: 'Contact', url: '/' },
+    ],
+  },
+];
