+++ src/data/social-links.js
@@ -1,0 +1,13 @@
+import {
+  FaFacebookSquare,
+  FaTwitter,
+  FaPinterest,
+  FaInstagram,
+} from 'react-icons/fa';
+
+export const socialLinks = [
+  { icon: <FaFacebookSquare />, url: '/', label: 'Facebook' },
+  { icon: <FaTwitter />, url: '/', label: 'Twitter' },
+  { icon: <FaPinterest />, url: '/', label: 'Pinterest' },
+  { icon: <FaInstagram />, url: '/', label: 'Instagram' },
+];
